requirements.txt
anyio==3.6.2
boto3==1.26.125
botocore==1.29.125
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
grpcio-tools==1.48.2
grpcio==1.54.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
jmespath==1.0.1
pip==22.3.1
protobuf==3.19.6
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-telegram-bot==20.2
requests==2.29.0
rfc3986==1.5.0
s3transfer==0.6.0
setuptools==65.5.1
six==1.16.0
sniffio==1.3.0
speechkit==2.2.2
urllib3==1.26.15
wheel==0.38.4

aioredis~=2.0.1
pyaspeller~=1.1.0
openai~=0.27.6